requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.5.1
billiard==3.3.0.23
bleach==1.5.0
celery==3.1.25
chardet==2.3.0
coreapi==2.3.1
coreschema==0.0.4
cssselect==1.0.1
Django==1.10.8
django-celery==3.2.1
django-contrib-comments==1.7.3
django-crispy-forms==1.6.1
django-extensions==1.8.1
django-filter==1.0.4
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
django-nested-formset==0.1.4
django-simple-captcha==0.5.3
djangorestframework==3.6.3
et-xmlfile==1.0.1
filebrowser-safe==0.4.6
future==0.16.0
grappelli-safe==0.4.5
html5lib==0.9999999
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
kombu==3.0.37
lxml==3.7.2
Markdown==2.6.8
MarkupSafe==1.0
Mezzanine==4.2.2
numpy==1.12.0
oauthlib==2.0.0
openpyxl==2.4.8
pandas==0.19.2
pdfminer==20140328
pdfquery==0.4.3
Pillow==3.4.2
psycopg2==2.6.2
Pygments==2.2.0
PyPDF2==1.26.0
pyquery==1.2.17
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
redis==2.10.6
requests==2.11.1
requests-oauthlib==0.7.0
roman==2.0.0
six==1.10.0
testfixtures==4.13.1
tzlocal==1.3
uk-postcode-utils==1.0
uritemplate==3.0.0
vine==1.1.3
xlrd==1.0.0
